Create Pickleball Store
Created a Pickleball Store and items ordered. Listed quantity of each item, what the cost is and what it can sell for (retail). Put it in order of cost so that I can see which effects the bottom line first.  Then totaled the quantity of how many items we will have ID's for.